made improvements to the chartes and carousel
