+++ src/components/PageHeader/NavBar/icons.tsx
@@ -1,0 +1,129 @@
+export const HomeIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="25"
+      height="24"
+      fill="none"
+      viewBox="0 0 25 24"
+    >
+      <path
+        fill="color(display-p3 0.2078 0.2078 0.4392)"
+        d="m20.54 6.82-5.76-4.03c-1.57-1.1-3.98-1.04-5.49.13L4.28 6.83c-1 .78-1.79 2.38-1.79 3.64v6.9c0 2.55 2.07 4.63 4.62 4.63h10.78c2.55 0 4.62-2.07 4.62-4.62V10.6c0-1.35-.87-3.01-1.97-3.78M13.25 18c0 .41-.34.75-.75.75s-.75-.34-.75-.75v-3c0-.41.34-.75.75-.75s.75.34.75.75z"
+      ></path>
+    </svg>
+  );
+};
+
+export const PortfolioIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="25"
+      height="24"
+      fill="none"
+      viewBox="0 0 25 24"
+    >
+      <path
+        stroke="color(display-p3 0.2078 0.2078 0.4392)"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="2"
+        d="m21.5 12-9 6-9-6m18 4-9 6-9-6m18-8-9 6-9-6 9-6z"
+      ></path>
+    </svg>
+  );
+};
+
+export const SearchIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      fill="none"
+      viewBox="0 0 20 20"
+    >
+      <path
+        fill="color(display-p3 0.2588 0.2588 0.5255)"
+        d="M8.334 2.083a6.25 6.25 0 0 1 4.97 10.04l3.952 3.954a.833.833 0 0 1-1.1 1.248l-.078-.07-3.954-3.952a6.25 6.25 0 1 1-3.79-11.22m0 1.667a4.583 4.583 0 1 0 0 9.166 4.583 4.583 0 0 0 0-9.166"
+      ></path>
+    </svg>
+  );
+};
+
+export const DollarIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      fill="#424286"
+      className="h-5 w-5"
+      viewBox="-1.7 0 20.4 20.4"
+    >
+      <path
+        id="SVGRepo_iconCarrier"
+        d="M16.417 10.283A7.917 7.917 0 1 1 8.5 2.366a7.916 7.916 0 0 1 7.917 7.917m-4.844 1.754a2.25 2.25 0 0 0-.556-1.477l-.001-.002a3 3 0 0 0-.835-.665l-.003-.002a3.5 3.5 0 0 0-.866-.313H9.31a3.8 3.8 0 0 0-.795-.083 3 3 0 0 1-.475-.037 1.8 1.8 0 0 1-.494-.158l-.002-.001a1.2 1.2 0 0 1-.371-.298L7.172 9a.73.73 0 0 1-.175-.44.75.75 0 0 1 .421-.63 2.16 2.16 0 0 1 1.11-.297 2.3 2.3 0 0 1 .391.066l.049.01a2.5 2.5 0 0 1 .473.166 1.3 1.3 0 0 1 .381.261.792.792 0 1 0 1.118-1.12 2.9 2.9 0 0 0-.85-.585 4 4 0 0 0-.785-.268h-.001l-.008-.002v-.786a.792.792 0 1 0-1.583 0v.763a3.6 3.6 0 0 0-1.14.454 2.33 2.33 0 0 0-1.159 1.967 2.3 2.3 0 0 0 .529 1.44 2.7 2.7 0 0 0 .894.717 3.3 3.3 0 0 0 .942.305 4.4 4.4 0 0 0 .736.059 2.2 2.2 0 0 1 .46.046 2 2 0 0 1 .467.168 1.4 1.4 0 0 1 .382.308.67.67 0 0 1 .165.436c0 .097 0 .324-.385.573a2.2 2.2 0 0 1-1.132.314 3.5 3.5 0 0 1-.494-.06 2.4 2.4 0 0 1-.459-.148h-.001a.95.95 0 0 1-.356-.274.792.792 0 1 0-1.197 1.037 2.5 2.5 0 0 0 .967.708 3.8 3.8 0 0 0 .774.237h.007v.783a.792.792 0 1 0 1.583 0v-.79a3.6 3.6 0 0 0 1.17-.479 2.22 2.22 0 0 0 1.107-1.9z"
+      ></path>
+    </svg>
+  );
+};
+
+export const EuroIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      fill="#424286"
+      className="mr-1.5 mt-0.5 size-4"
+      viewBox="0 0 32 32"
+    >
+      <path
+        id="SVGRepo_iconCarrier"
+        fillRule="evenodd"
+        d="M16 32C7.163 32 0 24.837 0 16S7.163 0 16 0s16 7.163 16 16-7.163 16-16 16M8 19.004h1.994Q11.649 24.999 17.9 25q2.377 0 4.1-.648v-2.373q-1.813.74-3.334.74-2.129 0-3.277-.815-1.15-.817-1.949-2.9h4.956l.8-2.003h-6.184q-.045-.567-.045-1.285 0-.624.056-.995h7.096l.811-2.003h-7.422q1.07-3.438 4.809-3.438a8.9 8.9 0 0 1 3.683.776V7.532A14 14 0 0 0 18.103 7q-6.24 0-8.02 5.718H8.812L8 14.721h1.69a13 13 0 0 0-.057 1.25q0 .498.034 1.03H8.81z"
+      ></path>
+    </svg>
+  );
+};
+
+export const PoundSterlingIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="20"
+      height="20"
+      fill="#424286"
+      className="mr-1.5 mt-0.5 size-4"
+      viewBox="0 0 32 32"
+    >
+      <path
+        id="SVGRepo_iconCarrier"
+        fillRule="evenodd"
+        d="M16 32C7.163 32 0 24.837 0 16S7.163 0 16 0s16 7.163 16 16-7.163 16-16 16m-4.913-17.185H9v2.637h2.087v4.462L9 22.523V24.5h13.947v-2.713h-8.585v-4.335h6.578v-2.637h-6.578V12.56c0-1.952 1.057-3.169 2.827-3.169 1.479 0 2.377.685 3.328 1.85L23 9.34c-1.268-1.598-2.906-2.84-5.838-2.84-3.856 0-6.075 2.307-6.075 5.983z"
+      ></path>
+    </svg>
+  );
+};
+
+export const DropdownDownIcon = () => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="12"
+      height="12"
+      fill="none"
+      viewBox="0 0 12 12"
+    >
+      <path
+        stroke="color(display-p3 0.2588 0.2588 0.5255)"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+        strokeWidth="1.5"
+        d="m9 4.5-3 3-3-3"
+      ></path>
+    </svg>
+  );
+};
