+++ src/components/ui/icons.tsx
@@ -1,42 +1,3 @@
-//stats bar icons
-export const CoinsIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-    >
-      <path
-        fill="#fff"
-        d="M8 1.333A6.67 6.67 0 0 0 1.335 8a6.67 6.67 0 0 0 6.667 6.666A6.67 6.67 0 0 0 14.667 8a6.67 6.67 0 0 0-6.666-6.667m2.354 7.487-3.453 2.393c-.507.353-.8.147-.654-.447l.634-2.56-1.1-.273c-.5-.12-.56-.467-.14-.76L9.094 4.78c.507-.354.8-.147.653.446l-.633 2.56 1.1.274c.5.126.56.466.14.76"
-      ></path>
-    </svg>
-  );
-};
-
-export const ExchangeIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-    >
-      <path
-        fill="#fff"
-        d="M10 15.167a.502.502 0 0 1-.427-.76l.7-1.167a.5.5 0 1 1 .86.513l-.18.3a4.18 4.18 0 0 0 3.22-4.06c0-.273.226-.5.5-.5.273 0 .5.227.5.5-.007 2.854-2.327 5.173-5.174 5.173M1.334 6.5a.504.504 0 0 1-.5-.5A5.175 5.175 0 0 1 6.001.833a.502.502 0 0 1 .426.76l-.7 1.167a.497.497 0 0 1-.686.166.497.497 0 0 1-.167-.686l.18-.3A4.174 4.174 0 0 0 1.834 6c0 .273-.227.5-.5.5M9.867 8.42v1.96c0 1.634-.653 2.287-2.286 2.287H5.62c-1.634 0-2.287-.653-2.287-2.286V8.42c0-1.634.653-2.287 2.287-2.287h1.96c1.633 0 2.286.653 2.286 2.286"
-      ></path>
-      <path
-        fill="#fff"
-        d="M10.38 3.333H8.42c-1.606 0-2.26.64-2.28 2.213h1.44c1.96 0 2.867.914 2.867 2.867v1.44c1.58-.02 2.214-.673 2.214-2.28V5.62c.006-1.634-.647-2.287-2.28-2.287"
-      ></path>
-    </svg>
-  );
-};
-
 export const DownIcon = () => {
   return (
     <svg
@@ -71,207 +32,67 @@ export const UpIcon = () => {
   );
 };
 
-export const BitcoinIcon = () => {
+// table loader svg
+export const FadeStaggerSquares = () => {
   return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 24 24"
-    >
-      <g clipPath="url(#clip0_587_10851)">
-        <path
-          fill="color(display-p3 0.9686 0.5765 0.1020)"
-          d="M24 12A12 12 0 1 1 .001 12 12 12 0 0 1 24 12"
-        ></path>
-        <path
-          fill="#fff"
-          fillRule="evenodd"
-          d="m8.529 5.48 2.758.738.616-2.297 1.379.374-.592 2.203 1.124.302.594-2.227 1.402.376-.604 2.238s2.291.507 2.83 2.37c.539 1.864-1.185 2.842-1.717 2.88 0 0 2.007 1.1 1.318 3.266s-2.806 2.552-5.033 2.056L12 20.081l-1.403-.376.617-2.285-1.113-.304-.617 2.302-1.392-.374.618-2.29-2.831-.764.713-1.584s.799.218 1.101.29.496-.242.582-.557c.085-.314 1.366-5.52 1.488-5.95.12-.428.072-.763-.437-.895-.508-.132-1.2-.338-1.2-.338zm2.781 6.895-.763 3.036s3.787 1.367 4.27-.556c.485-1.923-3.507-2.48-3.507-2.48m.352-1.44.75-2.781s3.241.58 2.842 2.128c-.4 1.548-2.31.954-3.592.653"
-          clipRule="evenodd"
-        ></path>
-      </g>
-      <defs>
-        <clipPath id="clip0_587_10851">
-          <path fill="#fff" d="M0 0h24v24H0z"></path>
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-export const EthereumIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="16"
-      height="16"
-      fill="none"
-      viewBox="0 0 16 16"
-    >
-      <g clipPath="url(#clip0_587_10854)">
-        <circle
-          cx="8"
-          cy="8"
-          r="8"
-          fill="color(display-p3 0.3843 0.4941 0.9176)"
-        ></circle>
-        <path
-          fill="color(display-p3 0.7529 0.7961 0.9647)"
-          d="m7.956 2.08-.08.269v7.83l.08.079 3.635-2.148z"
-        ></path>
-        <path fill="#fff" d="M7.955 2.08 4.32 8.11l3.635 2.148V2.079"></path>
-        <path
-          fill="color(display-p3 0.7529 0.7961 0.9647)"
-          d="m7.957 10.944-.045.055v2.789l.045.13 3.636-5.121z"
-        ></path>
-        <path fill="#fff" d="M7.955 13.918v-2.974L4.32 8.797z"></path>
-        <path
-          fill="color(display-p3 0.5059 0.5922 0.9333)"
-          d="M7.955 10.26 11.59 8.11 7.955 6.459z"
-        ></path>
-        <path
-          fill="color(display-p3 0.7529 0.7961 0.9647)"
-          d="m4.32 8.111 3.635 2.148v-3.8z"
-        ></path>
-      </g>
-      <defs>
-        <clipPath id="clip0_587_10854">
-          <path fill="#fff" d="M0 0h16v16H0z"></path>
-        </clipPath>
-      </defs>
-    </svg>
-  );
-};
-
-//nav links icons
-export const HomeIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="25"
-      height="24"
-      fill="none"
-      viewBox="0 0 25 24"
-    >
-      <path
-        fill="color(display-p3 0.2078 0.2078 0.4392)"
-        d="m20.54 6.82-5.76-4.03c-1.57-1.1-3.98-1.04-5.49.13L4.28 6.83c-1 .78-1.79 2.38-1.79 3.64v6.9c0 2.55 2.07 4.63 4.62 4.63h10.78c2.55 0 4.62-2.07 4.62-4.62V10.6c0-1.35-.87-3.01-1.97-3.78M13.25 18c0 .41-.34.75-.75.75s-.75-.34-.75-.75v-3c0-.41.34-.75.75-.75s.75.34.75.75z"
-      ></path>
-    </svg>
-  );
-};
-
-export const PortfolioIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="25"
-      height="24"
-      fill="none"
-      viewBox="0 0 25 24"
-    >
-      <path
-        stroke="color(display-p3 0.2078 0.2078 0.4392)"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="2"
-        d="m21.5 12-9 6-9-6m18 4-9 6-9-6m18-8-9 6-9-6 9-6z"
-      ></path>
-    </svg>
-  );
-};
-
-export const SearchIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      fill="none"
-      viewBox="0 0 20 20"
-    >
-      <path
-        fill="color(display-p3 0.2588 0.2588 0.5255)"
-        d="M8.334 2.083a6.25 6.25 0 0 1 4.97 10.04l3.952 3.954a.833.833 0 0 1-1.1 1.248l-.078-.07-3.954-3.952a6.25 6.25 0 1 1-3.79-11.22m0 1.667a4.583 4.583 0 1 0 0 9.166 4.583 4.583 0 0 0 0-9.166"
-      ></path>
-    </svg>
-  );
-};
-
-export const DollarIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      fill="#424286"
-      className="h-5 w-5"
-      viewBox="-1.7 0 20.4 20.4"
-    >
-      <path
-        id="SVGRepo_iconCarrier"
-        d="M16.417 10.283A7.917 7.917 0 1 1 8.5 2.366a7.916 7.916 0 0 1 7.917 7.917m-4.844 1.754a2.25 2.25 0 0 0-.556-1.477l-.001-.002a3 3 0 0 0-.835-.665l-.003-.002a3.5 3.5 0 0 0-.866-.313H9.31a3.8 3.8 0 0 0-.795-.083 3 3 0 0 1-.475-.037 1.8 1.8 0 0 1-.494-.158l-.002-.001a1.2 1.2 0 0 1-.371-.298L7.172 9a.73.73 0 0 1-.175-.44.75.75 0 0 1 .421-.63 2.16 2.16 0 0 1 1.11-.297 2.3 2.3 0 0 1 .391.066l.049.01a2.5 2.5 0 0 1 .473.166 1.3 1.3 0 0 1 .381.261.792.792 0 1 0 1.118-1.12 2.9 2.9 0 0 0-.85-.585 4 4 0 0 0-.785-.268h-.001l-.008-.002v-.786a.792.792 0 1 0-1.583 0v.763a3.6 3.6 0 0 0-1.14.454 2.33 2.33 0 0 0-1.159 1.967 2.3 2.3 0 0 0 .529 1.44 2.7 2.7 0 0 0 .894.717 3.3 3.3 0 0 0 .942.305 4.4 4.4 0 0 0 .736.059 2.2 2.2 0 0 1 .46.046 2 2 0 0 1 .467.168 1.4 1.4 0 0 1 .382.308.67.67 0 0 1 .165.436c0 .097 0 .324-.385.573a2.2 2.2 0 0 1-1.132.314 3.5 3.5 0 0 1-.494-.06 2.4 2.4 0 0 1-.459-.148h-.001a.95.95 0 0 1-.356-.274.792.792 0 1 0-1.197 1.037 2.5 2.5 0 0 0 .967.708 3.8 3.8 0 0 0 .774.237h.007v.783a.792.792 0 1 0 1.583 0v-.79a3.6 3.6 0 0 0 1.17-.479 2.22 2.22 0 0 0 1.107-1.9z"
-      ></path>
-    </svg>
-  );
-};
-
-export const EuroIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      fill="#424286"
-      className="mr-1.5 mt-0.5 size-4"
-      viewBox="0 0 32 32"
-    >
-      <path
-        id="SVGRepo_iconCarrier"
-        fillRule="evenodd"
-        d="M16 32C7.163 32 0 24.837 0 16S7.163 0 16 0s16 7.163 16 16-7.163 16-16 16M8 19.004h1.994Q11.649 24.999 17.9 25q2.377 0 4.1-.648v-2.373q-1.813.74-3.334.74-2.129 0-3.277-.815-1.15-.817-1.949-2.9h4.956l.8-2.003h-6.184q-.045-.567-.045-1.285 0-.624.056-.995h7.096l.811-2.003h-7.422q1.07-3.438 4.809-3.438a8.9 8.9 0 0 1 3.683.776V7.532A14 14 0 0 0 18.103 7q-6.24 0-8.02 5.718H8.812L8 14.721h1.69a13 13 0 0 0-.057 1.25q0 .498.034 1.03H8.81z"
-      ></path>
-    </svg>
-  );
-};
-
-export const PoundSterlingIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="20"
-      height="20"
-      fill="#424286"
-      className="mr-1.5 mt-0.5 size-4"
-      viewBox="0 0 32 32"
-    >
-      <path
-        id="SVGRepo_iconCarrier"
-        fillRule="evenodd"
-        d="M16 32C7.163 32 0 24.837 0 16S7.163 0 16 0s16 7.163 16 16-7.163 16-16 16m-4.913-17.185H9v2.637h2.087v4.462L9 22.523V24.5h13.947v-2.713h-8.585v-4.335h6.578v-2.637h-6.578V12.56c0-1.952 1.057-3.169 2.827-3.169 1.479 0 2.377.685 3.328 1.85L23 9.34c-1.268-1.598-2.906-2.84-5.838-2.84-3.856 0-6.075 2.307-6.075 5.983z"
-      ></path>
-    </svg>
-  );
-};
-
-export const DropdownDownIcon = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="12"
-      height="12"
-      fill="none"
-      viewBox="0 0 12 12"
-    >
-      <path
-        stroke="color(display-p3 0.2588 0.2588 0.5255)"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        strokeWidth="1.5"
-        d="m9 4.5-3 3-3-3"
-      ></path>
+    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 200 200">
+      <rect
+        fill="#A2A3ED"
+        stroke="#A2A3ED"
+        strokeWidth="6"
+        width="30"
+        height="30"
+        x="25"
+        y="85"
+      >
+        <animate
+          attributeName="opacity"
+          calcMode="spline"
+          dur="2"
+          values="1;0;1;"
+          keySplines=".5 0 .5 1;.5 0 .5 1"
+          repeatCount="indefinite"
+          begin="-.4"
+        ></animate>
+      </rect>
+      <rect
+        fill="#A2A3ED"
+        stroke="#A2A3ED"
+        strokeWidth="6"
+        width="30"
+        height="30"
+        x="85"
+        y="85"
+      >
+        <animate
+          attributeName="opacity"
+          calcMode="spline"
+          dur="2"
+          values="1;0;1;"
+          keySplines=".5 0 .5 1;.5 0 .5 1"
+          repeatCount="indefinite"
+          begin="-.2"
+        ></animate>
+      </rect>
+      <rect
+        fill="#A2A3ED"
+        stroke="#A2A3ED"
+        strokeWidth="6"
+        width="30"
+        height="30"
+        x="145"
+        y="85"
+      >
+        <animate
+          attributeName="opacity"
+          calcMode="spline"
+          dur="2"
+          values="1;0;1;"
+          keySplines=".5 0 .5 1;.5 0 .5 1"
+          repeatCount="indefinite"
+          begin="0"
+        ></animate>
+      </rect>
     </svg>
   );
 };
